Display kinch world rank

+++ server/src/main/java/org/worldcubeassociation/statistics/dto/sumofranks/SumOfRankEventDto.java
@@ -6,6 +6,6 @@ import org.worldcubeassociation.statistics.dto.EventDto;
 @Data
 public class SumOfRankEventDto {
     private EventDto event;
-    private Integer regionalRank;
+    private Float regionalRank;
     private Boolean completed;
 }
